backend/requirements.txt
Django>=4.0
djangorestframework>=3.14
Pillow 
reportlab
qrcode[pil] 
python-dotenv 